requirements.txt
flask
waitress
opencv-python
pillow
torchvision
numpy
ultralytics